- components > Start/Story/Game/Explore - screens
