+++ package.json
@@ -20,7 +20,7 @@
     "install": "^0.13.0",
     "npm": "^7.20.5",
     "react": "17.0.1",
-    "react-native": "0.64.2",
+    "react-native": "0.55.2",
     "react-native-camera": "git+https://git@github.com/react-native-community/react-native-camera.git",
     "react-native-gesture-handler": "^1.10.3",
     "react-native-linear-gradient": "^2.5.6",
