+++ src/components/Game/GameMain.js
@@ -7,23 +7,23 @@ import {
   Image,
   TouchableOpacity,
 } from 'react-native';
+import SpellingGame from './SpellingGame';
 
-function SpellingGame({navigation}) {
+const GameMain = ({navigation}) => {
   return (
     <View style={{flex: 1, zIndex: -1, display: 'flex'}}>
-      <TouchableOpacity
-        onPress={() => navigation.navigate('SpellingGameResult')}>
+      <TouchableOpacity onPress={() => navigation.navigate('SpellingGame')}>
         <Image
           style={{
             width: '100%',
             height: '100%',
           }}
-          source={require('../assets/gameWFriend.png')}
+          source={require('../../assets/game(ver3).png')}
           resizeMode="cover"
         />
       </TouchableOpacity>
     </View>
   );
-}
+};
 
-export default SpellingGame;
+export default GameMain;
