+++ src/components/Game/GameResult.js
