+++ src/components/Game/SpellingGameResult.js
@@ -7,23 +7,22 @@ import {
   Image,
   TouchableOpacity,
 } from 'react-native';
-import SpellingGame from './SpellingGame';
 
-const GameMain = ({navigation}) => {
+function SpellingGameResult({navigation}) {
   return (
     <View style={{flex: 1, zIndex: -1, display: 'flex'}}>
-      <TouchableOpacity onPress={() => navigation.navigate('SpellingGame')}>
+      <TouchableOpacity onPress={() => navigation.navigate('GameRank')}>
         <Image
           style={{
             width: '100%',
             height: '100%',
           }}
-          source={require('../assets/game(ver3).png')}
+          source={require('../../assets/gameWFriendResult.png')}
           resizeMode="cover"
         />
       </TouchableOpacity>
     </View>
   );
-};
+}
 
-export default GameMain;
+export default SpellingGameResult;
