+++ src/screens/GameScreen.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { createStackNavigator } from "@react-navigation/stack";
-import SpellingGame from './SpellingGame';
-import SpellingGameResult from './SpellingGameResult';
-import GameRank from './GameResult';
-import GameMain from './GameMain';
+import SpellingGame from '../components/Game/SpellingGame';
+import SpellingGameResult from '../components/Game/SpellingGameResult';
+import GameRank from '../components/Game/GameResult';
+import GameMain from '../components/Game/GameMain';
 
 const GameStack = createStackNavigator();
