+++ src/screens/NavTabs.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {StyleSheet, View, Text, Image} from 'react-native';
 import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
-import HomeScreen from './StoryMain';
+import HomeScreen from '../components/Story/StoryMain';
 import {icons, COLORS} from '../constants/index';
 import Settings from './Settings';
 import Explore from './ExploreScreen';
